Update dependencies and tools and re-organize
