+++ app/src/main/res/layout/fragment_details.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.DetailsFragment">
+    tools:context=".ui.fragment.DetailsFragment">
 
     <RelativeLayout
         android:id="@+id/fragment_details_map_wrapper"
