+++ app/src/main/res/layout/fragment_map.xml
@@ -19,6 +19,6 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             app:cameraZoom="13"
-            tools:context=".ui.MapFragment" />
+            tools:context=".ui.fragment.MapFragment" />
     </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
 </androidx.constraintlayout.widget.ConstraintLayout>
